Version 0.5.2

+++ strong-testimonials-dashboard.php
@@ -1,35 +1,53 @@
 <?php
 /**
  * Plugin Name: Strong Testimonials Dashboard
- * Version 0.4
+ * Version 0.5.2
  */
 
 // Exit if accessed directly
 if ( ! defined( 'ABSPATH' ) ) exit;
 
+/**
+ * Some style
+ */
+function strongdashboard_style() {
+	if ( class_exists( 'Kint' ) ) {
+		?>
+		<style>div[id^="strongdashboard_"] .kint footer { font-size: inherit; }</style>
+		<?php
+	}
+	else {
+		?>
+		<style>div[id^="strongdashboard_"] div pre { white-space: pre-wrap; }</style>
+		<?php
+	}
+}
+add_action( 'admin_head-index.php', 'strongdashboard_style' );
+
+
+/**
+ * Our array printer
+ *
+ * @param $option
+ */
 function strongdashboard_print_r( $option ) {
 	echo '<div>';
-	echo '<pre>' . print_r( $option, true ) . '</pre>';
+	if ( class_exists( 'Kint' ) ) {
+		echo d( $option );
+	} else {
+		echo '<pre>' . print_r( $option, true ) . '</pre>';
+	}
 	echo '</div>';
 }
 
-// force one-column dashboard
+/**
+ * Force one-column dashboard
+ */
 function strongdashboard_screen_layout_columns($columns) {
 	$columns['dashboard'] = 1;
 	return $columns;
 }
-add_filter('screen_layout_columns', 'strongdashboard_screen_layout_columns');
-
-
-function strongdashboard_screen_layout_dashboard() { return 1; }
-add_filter('get_user_option_screen_layout_dashboard', 'strongdashboard_screen_layout_dashboard');
-
-
-function strongdashboard_scripts() {
-	wp_enqueue_style( 'strongdashboard-style', plugins_url( 'style.css', __FILE__ ), false );
-}
-add_action( 'admin_enqueue_scripts', 'strongdashboard_scripts' );
-
+//add_filter('screen_layout_columns', 'strongdashboard_screen_layout_columns');
 
 /**
  * Add dashboard widgets.
@@ -44,55 +62,55 @@ function strongdashboard_add_dashboard_widgets() {
 
 	wp_add_dashboard_widget(
 		'strongdashboard_wpmtst_9',
-		'Base Forms',
+		'Strong Testimonials &bull; Base Forms',
 		'strongdashboard_wpmtst_option_9_function'
 	);
 
 	wp_add_dashboard_widget(
 		'strongdashboard_wpmtst_10',
-		'Custom Forms',
+		'Strong Testimonials &bull; Custom Forms',
 		'strongdashboard_wpmtst_option_10_function'
 	);
 
 	wp_add_dashboard_widget(
 		'strongdashboard_wpmtst_5',
-		'Strong Testimonials Views',
+		'Strong Testimonials &bull; Views',
 		'strongdashboard_wpmtst_option_5_function'
 	);
 
 	wp_add_dashboard_widget(
 		'strongdashboard_wpmtst_8',
-		'Strong Testimonials Templates',
+		'Strong Testimonials &bull; Templates',
 		'strongdashboard_wpmtst_option_8_function'
 	);
 
 	wp_add_dashboard_widget(
 		'strongdashboard_wpmtst_2',
-		'Strong Testimonials Fields',
+		'Strong Testimonials &bull; Fields',
 		'strongdashboard_wpmtst_option_2_function'
 	);
 
 	wp_add_dashboard_widget(
 		'strongdashboard_wpmtst_7',
-		'Strong Testimonials Default View',
+		'Strong Testimonials &bull; Default View',
 		'strongdashboard_wpmtst_option_7_function'
 	);
 
 	wp_add_dashboard_widget(
 		'strongdashboard_wpmtst_6',
-		'Strong Testimonials View Options',
+		'Strong Testimonials &bull; View Options',
 		'strongdashboard_wpmtst_option_6_function'
 	);
 
 	wp_add_dashboard_widget(
 		'strongdashboard_wpmtst_4',
-		'Strong Testimonials Form Options',
+		'Strong Testimonials &bull; Form Options',
 		'strongdashboard_wpmtst_option_4_function'
 	);
 
 	wp_add_dashboard_widget(
 		'strongdashboard_wpmtst_1',
-		'Strong Testimonials Options',
+		'Strong Testimonials &bull; Options',
 		'strongdashboard_wpmtst_option_1_function'
 	);
 
@@ -122,19 +140,17 @@ function strongdashboard_add_dashboard_widgets() {
 }
 add_action( 'wp_dashboard_setup', 'strongdashboard_add_dashboard_widgets', 20 );
 
+
 /**
- * ==============================
+ * ------------------------------
  * WIDGETS
- * ==============================
+ * ------------------------------
  */
 
 function strongdashboard_wpmtst_option_1_function() {
 	$options = get_option( 'wpmtst_options' );
 	if ( $options ) {
 		strongdashboard_print_r( $options );
-		if ( class_exists( 'Kint' ) ) {
-			d( $options );
-		}
 	}
 	else {
 		echo '<em>not found</em>';
@@ -145,9 +161,6 @@ function strongdashboard_wpmtst_option_2_function() {
 	$fields = get_option( 'wpmtst_fields' );
 	if ( $fields ) {
 		strongdashboard_print_r( $fields );
-		if ( class_exists( 'Kint' ) ) {
-			d( $fields );
-		}
 	}
 	else {
 		echo '<em>not found</em>';
@@ -158,9 +171,6 @@ function strongdashboard_wpmtst_option_4_function() {
 	$form_options = get_option( 'wpmtst_form_options' );
 	if ( $form_options ) {
 		strongdashboard_print_r( $form_options );
-		if ( class_exists( 'Kint' ) ) {
-			d( $form_options );
-		}
 	}
 	else {
 		echo '<em>not found</em>';
@@ -172,13 +182,10 @@ function strongdashboard_wpmtst_option_5_function() {
 		$views = wpmtst_get_views();
 		echo '<div>';
 		foreach ( $views as $key => $view ) {
-			echo '<p><b>' . $view['id'] . ' - ' . $view['name'] . '</b></p>';
-			echo '<pre>' . print_r( unserialize( $view['value'] ), true ) . '</pre>';
+			echo '<p style="font-size: 1.5em; margin-bottom: 0.5em; padding-top: 0.5em;">' . $view['id'] . ' - ' . $view['name'] . '</p>';
+			strongdashboard_print_r( unserialize( $view['value'] ) );
 		}
 		echo '</div>';
-		if ( class_exists( 'Kint' ) ) {
-			d( $views );
-		}
 	}
 	else {
 		echo '<em>not found</em>';
@@ -189,9 +196,6 @@ function strongdashboard_wpmtst_option_6_function() {
 	$view_options = get_option( 'wpmtst_view_options' );
 	if ( $view_options ) {
 		strongdashboard_print_r( $view_options );
-		if ( class_exists( 'Kint' ) ) {
-			d( $view_options );
-		}
 	}
 	else {
 		echo '<em>not found</em>';
@@ -202,9 +206,6 @@ function strongdashboard_wpmtst_option_7_function() {
 	$view_default = get_option( 'wpmtst_view_default' );
 	if ( $view_default ) {
 		strongdashboard_print_r( $view_default );
-		if ( class_exists( 'Kint' ) ) {
-			d( $view_default );
-		}
 	}
 	else {
 		echo '<em>not found</em>';
@@ -219,53 +220,14 @@ function strongdashboard_wpmtst_option_8_function() {
 
 	$strong_templates = new Strong_Templates();
 	$templates = $strong_templates->get_templates();
-	ob_start();
-	echo '<div>';
-	echo '<pre>'.print_r($templates,1).'</pre>';
-	echo '</div>';
-	$html = ob_get_contents();
-	ob_end_clean();
 
-	$html = str_replace(
-		array(
-			plugins_url(),
-			str_replace( '\\', '/', dirname( plugin_dir_path( __FILE__ ) ) ),
-
-			'strong-testimonials/templates',
-
-			get_template_directory_uri() . '/strong-testimonials',
-			str_replace( '\\', '/', get_template_directory() . '/strong-testimonials' ),
-
-			get_stylesheet_directory_uri() . '/strong-testimonials',
-			str_replace( '\\', '/', get_stylesheet_directory() . '/strong-testimonials' ),
-		),
-		array(
-			'[plugins-URI]',
-			'[plugins-path]',
-
-			'[plugin-template-dir]',
-
-			'[parent-theme-URI]',
-			'[parent-theme-path]',
-
-			'[child-theme-URI]',
-			'[child-theme-path]',
-		),
-		str_replace( '\\', '/', $html )
-	);
-
-	echo $html;
-
-	if ( class_exists( 'Kint' ) ) { d( $templates ); }
+	strongdashboard_print_r( $templates );
 }
 
 function strongdashboard_wpmtst_option_9_function() {
 	$base_forms = get_option( 'wpmtst_base_forms' );
 	if ( $base_forms ) {
 		strongdashboard_print_r( $base_forms );
-		if ( class_exists( 'Kint' ) ) {
-			d( $base_forms );
-		}
 	}
 	else {
 		echo '<em>not found</em>';
@@ -276,20 +238,8 @@ function strongdashboard_wpmtst_option_10_function() {
 	$custom_forms = get_option( 'wpmtst_custom_forms' );
 	if ( $custom_forms ) {
 		strongdashboard_print_r( $custom_forms );
-		if ( class_exists( 'Kint' ) ) {
-			d( $custom_forms );
-		}
 	}
 	else {
 		echo '<em>not found</em>';
 	}
 }
-
-/**
- * ==============================
- *  TEST FILTERS
- * ==============================
- */
-
-//add_filter( 'wpmtst_field_required_tag', '__return_false' );
-//add_filter( 'wpmtst_form_validation_script', '__return_false' );
